Replace threading with cron scheduler
Replaced the threading approach with the more high-level APScheduler library, which is now also in the requirements.txt.

The scraper will now actually run as well, although it will error due to the lack of a "data" directory and the scraper failing to check that it doesn't exist.

It's currently configured to run automatically every Monday at 4:00AM, in NY time.