Add parenthesis balance errors
In the case that parenthesis aren't balanced, errors are thrown, then caught in the top level to return no prerequisite for the course.

So far we've gotten lucky and they've been balanced every time, but it might be a good idea to create logic to clean this up instead of completely throwing out the bad data.