Resolve missing cookie issue
Small restructure of functions and an addition of clarifying docstrings.

The issue presented in the last commit turned out to be a result of a gap in my understanding of cookies; a JSESSIONID cookie is assigned on the first call to any SIS page.

I was inadvertently calling get_subjects() before attempting to read a cookie from get_cookie(), at which point the aiohttp session would already have a cookie and therefore would not receive another one from the server.

In addition, I did not know that the aiohttp session automatically stores cookies and sends them in subsequent requests. Therefore, there is no need to explicitly handle cookie logic.