Cache code mappings to disk
This ensures that names can reliably be converted to codes in the post-processing stage after scraping.

For example, there has been a case of a course referencing a prerequisite subject name that existed in a previous semester but not the current one.

Thus, caching these code mappings that contain data from all semesters will prevent cases like these.