Remove prereq parser temporarily
The scraper workflow is currently comprised of fetching raw names and codes from SIS and storing them in JSON, then post-processing that JSON to codify all names.

However, the prereq parser is currently incompatible with this system as it will attempt to directly spit out a 4-letter subject code for each course, so something like "Computer Science" gets truncated to "Comp".

The current prereq parser is essentially ripped from our old SIS scraper, which parsed prereqs from raw text data. However, the new SIS organizes prereqs into tables, so we will eventually rewrite the prereq parser to take advantage of this new structure.

We aren't using prereq data as of now, so this will be the last of the prereq parser for a while.